Move selectStackGroups to its own file
